Added ability to use custom meshes and configure them in blocks.conf file.
Meshes can be defined based on block ID and block data combination.

Meshes are either defined as the file name (within "conf" directory) or the file name and object name (within the file) separated by "#", eg: object.obj or object.obj#cake

Only quad based meshes are supported! Triangles are not implemented anywhere in the program so reading would be impossible for now.