Fully support n-sided polygons
Stopped writing normals again, unless they're explicitly included in the source model (TODO: implement transforms when normals are specified)

Support source models that omit normals and/or uvs

When referencing objects from .obj files (in blocks.conf), it's now possible to omit the object name, in which case the first object in the file is used. Useful for the common case where the .obj file only contains one object.